Add @@trigger-timing.
When running standalone @@triggers, runtest has to invoke the trigger process
before it invokes DTrace, in order to get a PID to pass DTrace.  Normally, it
runs a one-second sleep(1) before execing DTrace, to give DTrace time to start
and let it spy on all aspects of process startup: but some tests (notably USDT
tests) would like the trigger to start first.

So add a new @@trigger-timing option, whose value can be 'after' to indicate
that (as now) the trigger runs before DTrace, 'before' to run the trigger
first, or a time in seconds to put off running the trigger.  (The default
is 'after', to give the same behaviour as before.)