Link map iteration deadlock and dead-process detection fixes.
rd_loadobj_iter() and its associated consistency-enforcement machinery was
taking insufficient note of already-blocked and dead processes.

If a process is known to be not running, whether this is because it is stopped
at a breakpoint or not -- even if it is stopped because someone else has
SIGSTOPped it -- there is no point waiting for it to enter a consistent state.
Either it is in one already, or it will never be.  Equally, if it's known
to be dead after the initial Pwait() in rd_loadobj_iter(), there's no point
trying to iterate over it.

If a process with multiple lmids is discovered to be dead after we have waited
for it to be in a consistent state, we should always note that by returning -1,
rather than (as now) trying to Ptrace() it to stop it running.