Improve robustness of test/demo/act/ustack*.d.
Unfortunately, doing ustack()s on read() and pruning out all ld.so-related
syscalls was not sufficient: libc's internals vary too much, and spurious junk
left on the stack is often interpreted as other functions (that were, e.g.,
called beforehand and are now left in uninitialized bits of the stack frame).
Humans can interpret this fine, but diff has more problems with it.  Also,
other bits of libc sometimes do read()s, leading to entirely wrong stack frames
being checked.

So, instead of scanning read(), check lstat()/newlstat(), which are not issued
by anything in the guts of glibc in this case, and simply verify that particular
symbols are *present* in the stack trace, rather than checking what order they
are in or that there are no other symbols there.

This should be rather more robust.