Fix spurious double-breakpoint-removal bug.
rd_event_disable removes the r_brk breakpoint but does not reset the
rd_monitoring flag which indicates that this breakpoint is active.  If this
happens on process shutdown, a spurious removal can happen, leading to
a (harmless) warning.  If it were to happen at other times, a subsequent event
which needed to enable that breakpoint would fail to do so,
thinking (incorrectly) that it is already active.

Fix by resetting the flag.