Fix a variety of dph_lrucnt management errors.
Despite its name, dph_lrucnt is not a count of the number of items on the
dph_lrulist, but rather a count of the number of items on that list that are not
retired.  dt_proc_destroy was assuming the opposite, so if any processes were
ever retired, dph_lrucnt would fall to zero and trip an assertion failure during
the mass-dt_proc_destroy() at dtrace_close() time.

dt_proc_grab() contains a loop that scans the dph_lrulist for the least recently
used retired process, additionally verifying that it has a zero refcount.
Unfortunately it does this before boosting the refcount or unretiring the
process being grabbed, so an attempt to grab (and thus unretire) the oldest
retired process will end up retiring it immediately and then boosting its
refcount as though it were not retired, leading to an underflow of dph_lrucnt at
destroy time.