Fixes for test coverage analysis.
The rule for test coverage analysis is that only one program that uses test
coverage may run simultaneously (lest they corrupt their coverage outputs), and
no program that does not use test coverage may use any components that were
built with --coverage (or linking fails).  We had several bugs in this area:

 - drti.o and drti32.o were being built with coverage, so filter it out.

 - The various internals tests (libproc-lookup-by-name, libproc-pldd, and
   libproc-consistency) were being built without coverage, even though they need
   it because they link directly against DTrace internals, which are built with
   coverage.

In addition, the trigger-template foreach line was in the wrong place: it
must be *before* the per-file overrides, or it will override the overrides
and wipe many of them out.