Initial build system changes for SPARC64.
Rather than hardwiring include/i386 and libdtrace/i386 into include and source
paths, we compute an ARCHINC to include from the uname output and substitute it
into -I and *_SOURCES. To allow for source files that exist only on one
architecture, such as the disassembler in libdtrace/i386, we include an optional
Build file in the libdtrace platform directory and adjust libdtrace_SOURCES
there. (This inclusion must happen after the setting of libdtrace_DIR, since
$(current-dir) works by examining the list of included makefiles to determine
the "current" one.)

Import dt_isadep.c from the OpenSolaris DTrace tree, to ensure that all files
mentioned in the arch-independent Build files exist for SPARC too.  (It is
highly unlikely to compile without errors on SPARC at this stage.)