Fix lexer input bug breaking pragmata and comments at multiples of 8192 bytes.
The YY_INPUT() macro was testing YYSTATE in an attempt to detect pragmata and
comments at EOF.  This can never work, because YYSTATE is still recording the
state of the parser at the time YY_INPUT() is *called*: it is not updated
instantly by magic as soon as fread() is executed!  So if YY_INPUT is called
when the parser is inside a comment (say, at the 8192-byte boundary, the flex
read buffer size), and the file was less than 16384 bytes long, this code would
throw an error, even if the file ended 6000 chars later and not inside a comment
at all.

This would have worked in AT&T lex, since that did no buffering at all and
input() was called byte-by-byte, but in flex it cannot work.  This sort of thing
is what EOF rules were introduced for: so use those.

It was also failing to diagnose either error when parsing strings.

YY_INPUT() was also copying data from the yypcb->pcb_string byte-by-byte.  This
is ridiculous: it's hardly difficult to work out how much to copy and then
memcpy() it, which will almost always be both faster and simpler.

All fixed, tests added for both at-EOF detections and for this bug.
NEWS updated.

Jira: DTRACE-106