Work around GCC bug 36043.
This bug causes spurious reads of extra bytes when passing structures by
value.  Normally this is harmless, but when these reads run off the end
of an enclosing structure they may be problematic.  (Notably, unlike the
8-bytes-at-once optimized memcpy() et al, GCC does *not* guarantee that
such extra reads will not cross page boundaries, so they could cause
coredumps.)

A workaround is easy: add a bit of padding on the end of affected
enclosing structures.  So far, one such affected structure is known.