Check the right variable to detect stack read failure on x86-32.
On x86-32, we must read the stack to find the address of main().  We were
checking the wrong variable to check if that read had failed (which is highly
unlikely, but still).  This could have led to spurious failure if the random
garbage in this variable happened to look like a negative number.