Fix dtrace module requirements mess.
Drop the dtrace-kernel-interface / dtrace-modules-headers scheme, which
attempted to use Requires and BuildRequires to constrain the RPM solver.  This
cannot work: RPM follows Requires, it does not use them to constrain an overall
set of packages.  Thus you cannot use Requires to prevent installation of
packages providing conflicting features: only Conflicts can do that.

Fundamentally we cannot pull in the right dtrace-modules from the userspace
specfile at all: the right module to use changes as you change kernel.
The module already depends on the right kernel: so the problem is how to pull
it in.  Rather than trying to model it with RPM dependencies, which is hideous
and leads to situations in which the kernel depends on the dtrace module and
all users, even those not using DTrace, get the module lying around on disk,
have DTrace userspace solve the problem by taking special measures when the
attempt to open the dtrace device fails.

Actually cmd/dtrace.c *already* takes special measures: it retries with
DTRACE_O_NODEV flipped on, just in case it's been asked to do something like
dtrace -G which doesn't require access to the dtrace device at all.  So,
now, if that fails, we run the script 'load_dtrace_modules' in the dtrace
libdir (normally /usr/lib64/dtrace).

This can do whatever we like in future (loading lots of providers, etc), but for
now it just does a modprobe dtrace, then a quiet yum install of the modules and
another modprobe if that fails.  If the script succeeds, cmd/dtrace.c tries
reopening the dtrace device again, and only then gives up (well, actually it
does another DTRACE_O_NODEV run too, but we know that will fail so we can ignore
it).

The net effect is that a dtrace run with no module installed or modprobed will
install it and modprobe it if need be, if suitably privileged: otherwise the
on-screen output will be exactly as it was before.