Improve 'make help' output.
Rather than just providing help on options, we now provide help
on options, targets, and check targets.

In order for target help to be printed before check target help,
we move the miscellaneous targets in the top-level GNUmakefile
into a new Maketargets makefile, and add the miscellaneous target
help there.