Revive ustack nframes test: add a new test for multithreaded ustack.
The somewhat-bizarrely-named tst.spin.sh is a test for ustack()'s frame count
argument and for stability of ustack() across multiple invocations, grabbing
multiple ustacks from a process in a tight loop and then validating the output.
Now that profile providers work, this test works too.

GCC is throwing this off by optimizing out functions in utack-tst-spin.c: add
an appropriate CFLAGS for that.

Add a similar test, sans validation, to verify that multithreaded ustack()
works: but rip out the validation code. We'll use an expected-results file
and results postprocessor instead, when this test stops crashing the kernel.