Fix yydebug linking problem with bison >= 2.6.
bison 2.6's NEWS states:

*** New declarations (yacc.c, glr.c)

  The generated header now declares yydebug and yyparse.  Both honor
  --name-prefix=bar_, and yield

    int bar_parse (void);

  rather than

    #define yyparse bar_parse
    int yyparse (void);

  in order to facilitate the inclusion of several parser headers inside a
  single compilation unit.

As a side-effect of this change, apparently unnoticed, the declaration of
yydebug in the generated grammar has changed from a definition:

int yydebug;

to an extern declaration:

extern int yydebug;

Thus, we must now define a non-extern version of it somewhere.  We can safely
do this even on bison < 2.6 systems, and rely on the old Fortran common block
support to merge the two definitions.  (This is formally undefined in C, but GCC
has always supported it; even if GCC removes it in the future we will only need
it on build systems with an older Bison, which we know support this extension).

Orabug: 18551552