Rename DTRACE_PROC_NONINVASIVE to DTRACE_PROC_SHORTLIVED.
This flag is badly-named.  It's not a signal that 'this is to be a noninvasive
grab': that's merely the effect it currently has.  It's a signal that 'this grab
is short-lived: sacrifice accuracy if you wish'.  i.e., it's the grab you can
use if you know that the process won't be grabbed long enough for anyone to care
about the fact that you can't pick up dlopen()s as they happen.

This will be important soon, when this flag will also mean 'this process can
lose its ptrace(), if any, as it ages in the cache after release.'