Fix dt_symbol_by_addr() symbol-type scanning loop yet again.
Even after my fix of Nov 4 last year, this loop had two more bugs waiting in it.
Unlike the last bug, this only throws off symbol resolution if an ambiguous
symbol is detected, and there aren't many of those, only a few dozen in all the
kernel.  Unfortunately it also underruns the array...

There are two bugs.  Firstly, the 'symp' variable, which held the result of
dereferencing the pointer-to-a-pointer which we are walking over the array, did
make things clearer, but unfortunately it is also dangerous. The assignment of
'symp' must happen in either the loop increment or the loop body -- but the loop
test contains not only a check of the contents of the symbol cell referenced by
'symp' but *also* a check that *the pointer to it* lies within the bounds of the
per-module address-sorted array.  So if we look up the first symbol in a module,
our disambiguation check will walk off the start of it *and then dereference
that garbage* before noticing that in fact it is garbage and we shouldn't
dereference it.

This was fixed -- at the cost of making the function uglier -- by dropping the
symp variable entirely, and dereferencing the might-be-out-of-range pointer
explicitly everywhere we need to instead.

Secondly, our array back-walk was broken.  It was hopping over many elements,
not just one, and correspondingly shooting far off the start of the array fairly
often.  The cause is embarrassingly obvious: to decrement an array pointer, you
always use --, regardless of the type pointed to, and the compiler translates
that into a sizeof decrement.  You don't use sizeof yourself!

Orabug: 18603463