Add a version of Pread_scalar() which can optionally be quiet.
This version, for use when probing possibly-invalid addresses, can be asked to
not warn when read attempts fail.  (It still warns if the requested type
conversion is invalid.)

It is not always quiet, but rather quietness is controlled via a new parameter.
This lets us reimplement Pread_scalar() in terms of it, avoiding both code
duplication and massive code changes for the sake of a quietude option needed
only rarely.