Fix nested rd_ldso_consistent_begin()/end() calls.
The original intention of the (internal to rtld_db)
rd_ldso_consistent_begin()/end() functions was that they could be nested,
i.e. that you could call begin inside begin, and then two ends would be
necessary before the victim could enter an inconsistent state.

This was broken by a later bugfix, which tripped a failure or silent return on
entering begin() while halted at a breakpoint.  This means that multiple begins
might suddenly be balanced by only one end, iff the victim had hit a breakpoint
in the meantime.

Easily fixed by checking if we are in a begin/end pair at the top of the
function, and incrementing the count and returning early if so.  As a side
benefit this lets us ditch a bunch of checks of whether we were already in such
a pair (buggy checks, since not enough such checks were present, and it could
easly have deadlocked, had any existing code ever called begin nested inside
begin).

Needed for a later change which nests begins in just this way.