Do not allow rd_new() on a dead process.
If a process dies while under monitoring, dt_proc_control_cleanup() will free
the rtld_db handle via a call to Prelease(), which calls Psym_release(), which
deletes the rd handle and sets it to NULL.

This all happens under the dt_proc_lock, but when the control thread dies the
lock is released. Operations blocked on said lock can now proceed.  They don't
know the process is dead and can proceed into Plookup_by_addr() and other
functions, ultimately calling rd_new() (because there doesn't seem to be an
rtld_db).  This will often allow the creation to go through -- after all, we know
the address of r_debug and so on -- causing all sorts of trouble later.

Armour against this by preventing rd_new() on a dead process.

Orabug: 18550863