Avoid traversing a reallocated pointer to map_pmap.
The map_pmap, which describes an individual memory mapping belonging to a single
process, is part of the map_info structure, which is contained within the
mappings array in the ps_prochandle.  Pointers to instances of this structure
are extracted and retained within the prmap_files members of the map_files hash,
which contains all mappings corresponding to a single file, and identifies the
text and data mappings in particular.

Unfortunately, these pointers are extracted while the mappings are being read,
during which time the mappings array is being repeatedly reallocated.  Since the
map_pmap is contained within the mappings array, all those pointers are
potentially invalidated on every realloc(), leading to future lookups
potentially traversing freed memory and getting garbage, leading to subtle and
unpleasant symptoms such as a silent failure to correctly identify aliased
textual mappings to the executable, leading to spurious symbol resolution
failures much later on.  All these future lookups happen during the same
Pupdate_maps() call, but they are still traversals of freed memory and so can
still go wrong.

The somewhat invasive solution is to transform map_pmap into a separately
allocated structure, so that it is not reallocated during the repeated
realloc()s. The extra allocation per mapping might seem expensive, but in fact
it probably saves time for programs with many mappings (and is insignificant for
smaller ones): the potential cost of the repeated realloc()s rises exponentially
as the size of the mappings structure grows, while each malloc() is of roughly
constant cost regardless of their number.

Orabug: 18550863