Fix very unlikely buffer overrun in Pupdate_maps().
We look up the name of the executable in Pupdate_maps() so that we can compare
it with the name in /proc/$pid/maps (derived from the same in-kernel source) to
determine whether a mapping is the primary textual mapping of the executable or
not.  The symlink can, of course, contain an absolute filename no longer than
PATH_MAX.  Unfortunately, when this symlink is near its maximum length and
refers to a deleted file, the string " (deleted)" is appended to it, which can
push it over PATH_MAX, leading to an up-to-ten-byte buffer overrun.

As an aside, this whole string-comparison business is very far from ideal.  In
future we should stat() /proc/$pid/exe and compare its (dev, ino) pair to the
pair wired into /proc/$pid/maps.  (Very unusually, this symlink can be traversed
even when "broken" by pointing to a deleted file, so we can still get that
information out of it.)

Orabug: 18550863