New 'cppargs' option to pass arbitrary arguments to CPP.
DTrace often runs D scripts through /usr/bin/cpp.  It passes a number of
hardwired command-line arguments to it, and has a few options to pass more
specific arguments (-D, for instance), but there is no way to pass an arbitrary
argument.

The testsuite needs to pass a bunch of -f options to make GCC 4.8 produce error
output more like that of GCC 4.4 for testing purposes, so we need an option to
do so.

So introduce such an option, 'cppargs'.  Vaguely shell-like word splitting is
performed as on the arguments to the DTrace -c option, so you can pass multiple
arguments (but not quoted arguments, as no quote processing performed).

Orabug: 19054052