Fix the types/unloaded_var test.
This test relies on the existence of at least one module that is installed but
is unlikely to be loaded.  We used can.ko, but that's gone as of the OL7 config.

hysdn.ko is present on both: it's a card that as far as I can see hasn't been
made in years, and, well, it's ISDN.  It screams 'obsolete' to me, so let's try
using it for this purpose for now.

Orabug: 19054052