Introduce Pelf64(); use it in userspace tracepointing code.
The dt_pid code attempts to differentiate between 64-bit and 32-bit
processes vai Pstatus()->pr_dmodel.  This has not worked for a long
time and was only missed because it was wrapped in

Remove that #define and rewire to use the new Pelf64() instead, which consults
the internal elf64 flag, ultimately extracted from the ELF class in the ELF
header.

No visible effect (yet), since the pid provider is not yet ported.