Comment fixes, and renaming of a function.
dt_proc_attach() is a really bad name for this function, since when it is
called, we are already attached.  It's attaching a breakpoint *to* the process,
not attaching *us* to the process, so rename it to dt_proc_attach_break().

(Apologies for the misindentation here: it will be fixed up in the next commit.)